Split out evaluation model classes into purpose based files and added basic template constructor logic
