+++ python/lib/core/dmod/core/common/__init__.py
@@ -11,6 +11,7 @@ from .helper_functions import truncate
 from .helper_functions import is_true
 from .helper_functions import to_json
 from .helper_functions import order_dictionary
+from .helper_functions import find
 from .helper_functions import generate_identifier
 from .helper_functions import generate_key
 from .tasks import wait_on_task
