+++ python/lib/evaluations/dmod/evaluations/specification/__init__.py
@@ -1,16 +1,20 @@
-from .model import EvaluationSpecification
-from .model import ThresholdDefinition
-from .model import ThresholdSpecification
-from .model import SchemeSpecification
-from .model import LocationSpecification
-from .model import DataSourceSpecification
-from .model import MetricSpecification
-from .model import FieldMappingSpecification
-from .model import ValueSelector
-from .model import CrosswalkSpecification
-from .model import BackendSpecification
-from .model import AssociatedField
-from .model import UnitDefinition
-from .model import ThresholdApplicationRules
-from .model import EvaluationResults
-from .model import LoaderSpecification
+from .template import TemplateManager
+from .template import TemplateDetails
+from .base import Specification
+from .base import TemplatedSpecification
+from .evaluation import EvaluationSpecification
+from .threshold import ThresholdDefinition
+from .threshold import ThresholdSpecification
+from .scoring import SchemeSpecification
+from .locations import LocationSpecification
+from .data import DataSourceSpecification
+from .scoring import MetricSpecification
+from .fields import FieldMappingSpecification
+from .fields import ValueSelector
+from .locations import CrosswalkSpecification
+from .backend import BackendSpecification
+from .fields import AssociatedField
+from .unit import UnitDefinition
+from .threshold import ThresholdApplicationRules
+from .evaluation import EvaluationResults
+from .backend import LoaderSpecification
